Factor our transactionConstructionSelectors into their own file
