+++ background/redux-slices/transaction-construction.ts
@@ -1,4 +1,4 @@
-import { createSlice, createSelector } from "@reduxjs/toolkit"
+import { createSlice } from "@reduxjs/toolkit"
 import Emittery from "emittery"
 import { FORK } from "../constants"
 import {
@@ -21,7 +21,6 @@ import {
   EnrichedEVMTransactionSignatureRequest,
 } from "../services/enrichment"
 import { SigningMethod } from "../utils/signing"
-import { UIState } from "./ui"
 
 import { createBackgroundAsyncThunk } from "./utils"
 
@@ -293,76 +292,4 @@ export const rejectTransactionSignature = createBackgroundAsyncThunk(
       )
     )
   }
-)
-
-export const selectDefaultNetworkFeeSettings = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction,
-  (state: { ui: UIState }) => state.ui.selectedAccount.network,
-  (transactionConstruction, network): NetworkFeeSettings => {
-    const selectedFeesPerGas =
-      transactionConstruction.estimatedFeesPerGas?.[network.chainID]?.[
-        transactionConstruction.feeTypeSelected
-      ]
-    return {
-      feeType: transactionConstruction.feeTypeSelected,
-      gasLimit: undefined,
-      suggestedGasLimit: transactionConstruction.transactionRequest?.gasLimit,
-      values: {
-        maxFeePerGas: selectedFeesPerGas?.maxFeePerGas ?? 0n,
-        maxPriorityFeePerGas: selectedFeesPerGas?.maxPriorityFeePerGas ?? 0n,
-      },
-    }
-  }
-)
-
-export const selectEstimatedFeesPerGas = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.estimatedFeesPerGas,
-  (state: { ui: UIState }) => state.ui.selectedAccount.network,
-  (gasData, selectedNetwork) => gasData[selectedNetwork.chainID]
-)
-
-export const selectFeeType = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.feeTypeSelected,
-  (feeTypeChosen) => feeTypeChosen
-)
-
-export const selectLastGasEstimatesRefreshTime = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.lastGasEstimatesRefreshed,
-  (updateTime) => updateTime
-)
-
-export const selectTransactionData = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.transactionRequest,
-  (transactionRequestData) => transactionRequestData
-)
-
-export const selectIsTransactionPendingSignature = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.status,
-  (status) => status === "loaded" || status === "pending"
-)
-
-export const selectIsTransactionLoaded = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.status,
-  (status) => status === "loaded"
-)
-
-export const selectIsTransactionSigned = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction.status,
-  (status) => status === "signed"
-)
-
-export const selectCurrentlyChosenNetworkFees = createSelector(
-  (state: { transactionConstruction: TransactionConstruction }) =>
-    state.transactionConstruction?.estimatedFeesPerGas?.[
-      state.transactionConstruction.feeTypeSelected
-    ],
-  (feeData) => feeData
 )
