+++ ui/components/Swap/SwapTransactionSettingsChooser.tsx
@@ -1,9 +1,11 @@
 import {
   NetworkFeeSettings,
-  selectDefaultNetworkFeeSettings,
-  selectEstimatedFeesPerGas,
   setFeeType,
 } from "@tallyho/tally-background/redux-slices/transaction-construction"
+import {
+  selectDefaultNetworkFeeSettings,
+  selectEstimatedFeesPerGas,
+} from "@tallyho/tally-background/redux-slices/selectors/transactionConstructionSelectors"
 
 import React, { ReactElement, useState } from "react"
 import { SWAP_FEE } from "@tallyho/tally-background/redux-slices/0x-swap"
