+++ ui/pages/Send.tsx
@@ -6,9 +6,9 @@ import {
 } from "@tallyho/tally-background/redux-slices/selectors"
 import {
   NetworkFeeSettings,
-  selectEstimatedFeesPerGas,
   setFeeType,
 } from "@tallyho/tally-background/redux-slices/transaction-construction"
+import { selectEstimatedFeesPerGas } from "@tallyho/tally-background/redux-slices/selectors/transactionConstructionSelectors"
 import {
   FungibleAsset,
   isFungibleAssetAmount,
